pyproject.toml
[tool.poetry]
name = "discord-dict"
version = "0.1.0"
description = ""
authors = ["Kevin Bernfeld <kcbernfeld@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
requests = "^2.27.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.942"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
